move `puppet apply` to script and confirm puppet exit status; add webdev_postprocessor_dryrun option

+++ centos-7/bin/webdev_postprocessor.sh
@@ -2,6 +2,7 @@
 
 set -e
 
+WEBDEV_POSTPROCESSOR_DRYRUN=${WEBDEV_POSTPROCESSOR_DRYRUN:-0}
 CHANGELOG="${CHANGELOG:-routine update}"
 LOCAL_WEBDEV_JSON='builds/vagrant/webdev.json'
 LOCAL_BOX='builds/vagrant/virtualbox/centos-7-64-virtualbox-web.box'
@@ -79,7 +80,14 @@ jq \
 echo 'EBRC: Create a directory on our vagrant box server'
 ssh "${BOX_SERVER}" "mkdir -p ${BOX_SERVER_PATH}/webdev/${VERSION}"
 
-echo 'EBRC: Upload the box to versioned directory, and json file to web root directory.'
-rsync -qaPv "$LOCAL_WEBDEV_JSON" "${BOX_SERVER}:${BOX_SERVER_PATH}/"
-rsync -qaPv builds/vagrant/virtualbox/centos-7-64-virtualbox-web.box \
-  "${BOX_SERVER}:${BOX_SERVER_PATH}/webdev/${VERSION}"
+if [[ "$WEBDEV_POSTPROCESSOR_DRYRUN" -eq 0 ]]; then
+  echo 'EBRC: Upload the box to versioned directory, and json file to web root directory.'
+  rsync -qaPv "$LOCAL_WEBDEV_JSON" "${BOX_SERVER}:${BOX_SERVER_PATH}/"
+  rsync -qaPv builds/vagrant/virtualbox/centos-7-64-virtualbox-web.box \
+    "${BOX_SERVER}:${BOX_SERVER_PATH}/webdev/${VERSION}"
+else
+  echo 'EBRC: (DRY-RUN:) Upload the box to versioned directory, and json file to web root directory.'
+  echo "(DRY-RUN:) rsync -qaPv $LOCAL_WEBDEV_JSON" "${BOX_SERVER}:${BOX_SERVER_PATH}/"
+  echo "(DRY-RUN:) rsync -qaPv builds/vagrant/virtualbox/centos-7-64-virtualbox-web.box \
+    ${BOX_SERVER}:${BOX_SERVER_PATH}/webdev/${VERSION}"
+fi
